patches/l_sar/config: remove old config parameters
Since LXC is stablised to v4.X now and gentoo has dropped support for
older versions, we have to stick to it.
But v4.0 strips off crucial parameters like altering cgroup limits for
the container. This obviously will have to be dealt in some other way
now, but from what I have tried, experimenting with lxc config is not
the way to do it.

Signed-off-by: Gunwant Jain <mail@wantguns.dev>